// src/status/status.enum.ts
export enum Status_ENUM {
    AVAILABLE = 'AVAILABLE',
    UNAVAILABLE = 'UNAVAILABLE',
    ACTIVE = 'ACTIVE',
    INACTIVE = 'INACTIVE',
}
export enum Status_ENUM_DISCOUNT {
    AVAILABLE = 'AVAILABLE',
    UNAVAILABLE = 'UNAVAILABLE',
    NOSTATUS = 'NO',
}

export enum Status_Ticket {
    SUCCESS = 'SUCCESS',
    PROCESS = 'PROCESS',
    DONE = 'DONE',
    CANCEL = 'CANCEL',
}